Abstract ConcludableResolver and RetrievableResolver with Coordinator.
